refactoring (no change in functionality): extract publishing logic in new supporting class PlatformAgentStreamPublisher
